[cms-core] Restructure files and clean up
