+++ cms/core/src/infrastructure/rendering/content-area/content-area.ts
