+++ cms/core/src/public-api.ts
@@ -1,49 +1,50 @@
 /*
  * Public API Surface of core
  */
-
-export * from './utils/appInjector';
-export * from './utils/undetected.event';
-export * from './utils/route-reuse-strategy';
-
+export * from './angular-cms.module';
 export * from './bases/cms-component';
 export { CmsProperty } from './bases/cms-property';
-export { CmsPropertyFactoryResolver, CmsPropertyFactory, PROPERTY_PROVIDERS_TOKEN, getCmsPropertyFactory } from './bases/cms-property.factory';
-export * from './bases/content-data';
+export { CmsPropertyFactory, CmsPropertyFactoryResolver, getCmsPropertyFactory, PROPERTY_PROVIDERS_TOKEN } from './bases/cms-property.factory';
 export * from './bases/selection-factory';
 
-export * from './render/cms-content';
+export * from './cms';
 export * from './constants';
-export * from './types';
-export * from './types/module-config';
-export * from './types/content-type';
-export * from './types/ui-hint'
+export * from './core.module';
 
-export * from './decorators/metadata-key';
 export * from './decorators/content-type.decorator';
+export * from './decorators/metadata-key';
 export * from './decorators/property.decorator';
 export * from './decorators/validate.decorator';
 
-export * from './directives/insert-point.directive';
-export * from './render/content-area/content-area.directive';
-export * from './render/property/cms-property.directive';
-export * from './render/property/property-render.factory';
+export { clone, generateUUID, sortTabByTitle } from './helpers/common';
 
-export * from './models/block.model';
-export * from './models/content.model';
-export * from './models/media.model';
-export * from './models/page.model';
 
-export { sortTabByTitle, clone, generateUUID } from './helpers/common';
+export * from './infrastructure/browser/browser-location.service';
+export * from './infrastructure/browser/browser-storage.service';
+export * from './infrastructure/config/config.factory';
+export * from './infrastructure/config/config.service';
+
+export * from './infrastructure/rendering/cms-content';
+export * from './infrastructure/rendering/content-area/content-area.directive';
+export * from './infrastructure/rendering/insert-point.directive';
+export * from './infrastructure/rendering/property/cms-property.directive';
+export * from './infrastructure/rendering/property/property-render.factory';
 
 export * from './services/content-type.service';
-export * from './services/block.service';
-export * from './services/page.service';
-export * from './services/media.service';
-export * from './services/browser-location.service';
-export * from './services/browser-storage.service';
-export * from './services/config.service';
+export * from './services/content/block.service';
+export * from './services/content/media.service';
+export * from './services/content/models/block.model';
+export * from './services/content/models/content-data';
+export * from './services/content/models/content.model';
+export * from './services/content/models/media.model';
+export * from './services/content/models/page.model';
+export * from './services/content/page.service';
 
-export * from './cms';
-export * from './core.module';
-export * from './angular-cms.module';
+export * from './types';
+export * from './types/content-type';
+export * from './types/module-config';
+export * from './types/ui-hint';
+
+export * from './utils/appInjector';
+export * from './utils/route-reuse-strategy';
+export * from './utils/undetected.event';
