+++ cms/core/src/services/base.service.ts
@@ -1,0 +1,13 @@
+
+import { HttpClient } from '@angular/common/http';
+import { AppInjector } from '../utils/appInjector';
+import { ConfigService } from '../infrastructure/config/config.service';
+
+export abstract class BaseService {
+    protected baseApiUrl: string = AppInjector ? AppInjector.get(ConfigService).baseApiUrl : '';
+    protected httpClient: HttpClient;
+    protected abstract apiUrl: string;
+    constructor(httpClient: HttpClient) {
+        this.httpClient = httpClient;
+    }
+}
