+++ cms/core/src/services/content/models/content.model.ts
