+++ cms/core/src/services/content/models/media.model.ts
